course_database/requirements.txt
requests
beautifulsoup4
openai
tqdm
aiohttp
python-dotenv
tenacity
